Create restaurant, add menu item and place order

+++ src/main/java/dev/lekha/factory/MenuItemFactory.java
@@ -5,7 +5,7 @@ import dev.lekha.model.MenuItem;
 import java.util.UUID;
 
 public class MenuItemFactory {
-    public MenuItem defaultMenuItem(final String name, final double price) {
+    public MenuItem defaultMenuItem(final String name, final float price) {
         String menuId = UUID.randomUUID().toString();
         return new MenuItem(menuId, name, price);
     }
