+++ src/main/java/dev/lekha/model/MenuItem.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class MenuItem {
     private final String id;
     private final String name;
-    private final double price;
+    private final float price;
 
-    public MenuItem(String id, String name, double price) {
+    public MenuItem(String id, String name, float price) {
         this.id = id;
         this.name = name;
         this.price = price;
